Реализована логика покупки валюты

+++ src/main/java/com/demo/exceptions/NotEnoughFundsException.java
@@ -1,8 +1,10 @@
 package com.demo.exceptions;
 
+import java.math.BigDecimal;
+
 public class NotEnoughFundsException extends BankException {
 
-    public NotEnoughFundsException(Integer id, Double balance, long amount, String message) {
+    public NotEnoughFundsException(Integer id, BigDecimal balance, BigDecimal amount, String message) {
         super(message + " id : " + id + " balance : " + " amount : " + amount);
     }
